requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlalchemy
alembic
psycopg2-binary
redis
celery
pydantic
pydantic-settings
openai
langchain
langchain-openai
python-dotenv
